+++ APP/servo/servo.c
@@ -99,7 +99,7 @@ void toCenter(void){
 
 }
 
-void changeAngle(u8 angle){
+void changeAngle(short angle){
 	
 	float temp;
